Closes: #124, #125
Finalize statements if an error occurs during BEGIN, COMMIT or ROLLBACK. This is based on the idea of the finalized_statements.patch from ticket:126.

Make sure that the result of the calculated timout doesn't overflow. (ticket:124)

Get rid of the Connection* cast in _sqlite_step_with_busyhandler. (ticket:125).

Implmented a new module-level function to check if a string is a complete SQL statement.